adicionando endpoints para deletar assessments

+++ src/main/java/com/secretaria/secretaria/dto/grades/SendGradesDTO.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class SendGradesDTO {
-    private final Long assessmentId;
+    private final Integer assessmentId;
     private final Double grade;
     private final Integer studentId;
     private final Integer teacherId;
